requirements.txt
daphne~=4.2.1
Django~=6.0
djangorestframework~=3.16.1
django-cors-headers~=4.9.0
django-filter~=25.2
asgiref~=3.11.0
channels~=4.3.2
dotenv~=0.9.9
psycopg2-binary~=2.9.11
python-dotenv~=1.2.1
